refactor (SectionService) : SectionService 기능을 LineService 로 이동
